Schema objects restructure
