+++ openapi_core/extensions/__init__.py
@@ -1,0 +1,1 @@
+"""OpenAPI extensions package"""
