+++ openapi_core/extensions/models/__init__.py
@@ -1,0 +1,1 @@
+"""OpenAPI X-Model extension package"""
