+++ openapi_core/schema/infos/models.py
@@ -1,0 +1,8 @@
+"""OpenAPI core infos models module"""
+
+
+class Info(object):
+
+    def __init__(self, title, version):
+        self.title = title
+        self.version = version
