+++ openapi_core/schema/operations/generators.py
@@ -1,50 +1,14 @@
 # -*- coding: utf-8 -*-
-"""OpenAPI core operations module"""
-import logging
+"""OpenAPI core operations models module"""
 from functools import lru_cache
 
 from six import iteritems
 from openapi_spec_validator.validators import PathItemValidator
 
-from openapi_core.exceptions import InvalidResponse
-from openapi_core.parameters import ParametersGenerator
-from openapi_core.request_bodies import RequestBodyFactory
-from openapi_core.responses import ResponsesGenerator
-
-log = logging.getLogger(__name__)
-
-
-class Operation(object):
-    """Represents an OpenAPI Operation."""
-
-    def __init__(
-            self, http_method, path_name, responses, parameters,
-            request_body=None, deprecated=False, operation_id=None):
-        self.http_method = http_method
-        self.path_name = path_name
-        self.responses = dict(responses)
-        self.parameters = dict(parameters)
-        self.request_body = request_body
-        self.deprecated = deprecated
-        self.operation_id = operation_id
-
-    def __getitem__(self, name):
-        return self.parameters[name]
-
-    def get_response(self, http_status='default'):
-        try:
-            return self.responses[http_status]
-        except KeyError:
-            # try range
-            http_status_range = '{0}XX'.format(http_status[0])
-            if http_status_range in self.responses:
-                return self.responses[http_status_range]
-
-            if 'default' not in self.responses:
-                raise InvalidResponse(
-                    "Unknown response http status {0}".format(http_status))
-
-            return self.responses['default']
+from openapi_core.schema.operations.models import Operation
+from openapi_core.schema.parameters.generators import ParametersGenerator
+from openapi_core.schema.request_bodies.factories import RequestBodyFactory
+from openapi_core.schema.responses.generators import ResponsesGenerator
 
 
 class OperationsGenerator(object):
