+++ openapi_core/schema/parameters/enums.py
@@ -1,3 +1,4 @@
+"""OpenAPI core parameters enums module"""
 from enum import Enum
 
 
@@ -22,27 +23,3 @@ class ParameterStyle(Enum):
     SPACE_DELIMITED = 'spaceDelimited'
     PIPE_DELIMITED = 'pipeDelimited'
     DEEP_OBJECT = 'deepObject'
-
-
-class SchemaType(Enum):
-
-    INTEGER = 'integer'
-    NUMBER = 'number'
-    STRING = 'string'
-    BOOLEAN = 'boolean'
-    ARRAY = 'array'
-    OBJECT = 'object'
-
-
-class SchemaFormat(Enum):
-
-    NONE = None
-    INT32 = 'int32'
-    INT64 = 'int64'
-    FLOAT = 'float'
-    DOUBLE = 'double'
-    BYTE = 'byte'
-    BINARY = 'binary'
-    DATE = 'date'
-    DATETIME = 'date-time'
-    PASSWORD = 'password'
