+++ openapi_core/schema/schemas/enums.py
@@ -1,0 +1,26 @@
+"""OpenAPI core schemas enums module"""
+from enum import Enum
+
+
+class SchemaType(Enum):
+
+    INTEGER = 'integer'
+    NUMBER = 'number'
+    STRING = 'string'
+    BOOLEAN = 'boolean'
+    ARRAY = 'array'
+    OBJECT = 'object'
+
+
+class SchemaFormat(Enum):
+
+    NONE = None
+    INT32 = 'int32'
+    INT64 = 'int64'
+    FLOAT = 'float'
+    DOUBLE = 'double'
+    BYTE = 'byte'
+    BINARY = 'binary'
+    DATE = 'date'
+    DATETIME = 'date-time'
+    PASSWORD = 'password'
