+++ tests/unit/test_specs.py
@@ -2,8 +2,8 @@ import mock
 import pytest
 
 from openapi_core.exceptions import InvalidOperation
-from openapi_core.paths import Path
-from openapi_core.specs import Spec
+from openapi_core.schema.paths.models import Path
+from openapi_core.schema.specs.models import Spec
 
 
 class TestSpecs(object):
